pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb==0.10.0",
    "fastparquet==2024.2.0",
    "pandas==2.2.0",
    "psycopg2-binary==2.9.9",
    "pyarrow==15.0.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "sqlalchemy==2.0.25",
]
